requirements.txt
tensorflow==2.3.0
NBT==1.5.0
frozendict==1.2
anvil-parser==0.9.0
webcolors==1.11.1
loguru==0.5.3
tensorflow-graphics==2021.12.3